# application/api-spec/src/main/resources/openapi/skill-market-api.yaml
openapi: 3.0.3
info:
  title: Skill Market API
  version: 1.0.0
  description: API for freelancer and client project management.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/v1/projects:
    get:
      tags:
        - Projects
      operationId: getAllProjects
      summary: Get marketplace projects
      description: Returns projects that are visible in the marketplace.
      responses:
        '200':
          description: Projects loaded successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectResponse'
        '401':
          description: Missing or invalid access token.
        '403':
          description: Access token does not have required role.
    post:
      tags:
        - Projects
      operationId: createProject
      summary: Create a new project
      description: Creates a new project in OPEN state so freelancers can react to it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid access token.
        '403':
          description: Access token does not have required role.
        '409':
          description: Project with the same title already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{projectId}:
    put:
      tags:
        - Projects
      operationId: updateProject
      summary: Update project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: User is not allowed to manage this project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Project can no longer be edited.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Projects
      operationId: cancelProject
      summary: Cancel project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '204':
          description: Project cancelled successfully.
        '403':
          description: User is not allowed to manage this project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Project can no longer be cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/freelancers:
    get:
      tags:
        - Freelancers
      operationId: searchFreelancers
      summary: Search freelancers
      description: Returns freelancer profiles and optionally filters them by a skill.
      parameters:
        - name: skill
          in: query
          required: false
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Freelancers loaded successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FreelancerResponse'
        '401':
          description: Missing or invalid access token.
        '403':
          description: Access token does not have required role.
  /api/v1/profiles/{userId}:
    get:
      tags:
        - Profiles
      operationId: getProfile
      summary: Get user profile
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Profile loaded successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '404':
          description: User profile not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Profiles
      operationId: updateProfile
      summary: Update user profile
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '400':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User profile not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{projectId}/offers:
    get:
      tags:
        - Offers
      operationId: getProjectOffers
      summary: List project offers
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project offers loaded successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OfferResponse'
        '403':
          description: User is not allowed to see project offers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Offers
      operationId: createOffer
      summary: Submit offer to project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOfferRequest'
      responses:
        '201':
          description: Offer created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferResponse'
        '400':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project or freelancer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Project does not accept offers or user is not a freelancer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{projectId}/offers/{offerId}:
    delete:
      tags:
        - Offers
      operationId: deleteOffer
      summary: Cancel submitted offer
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/OfferId'
      responses:
        '204':
          description: Offer deleted successfully.
        '404':
          description: Project or offer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Offer can no longer be cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{projectId}/offers/{offerId}/accept:
    post:
      tags:
        - Offers
      operationId: acceptOffer
      summary: Accept offer for project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/OfferId'
      responses:
        '200':
          description: Offer accepted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferResponse'
        '403':
          description: User is not allowed to manage this project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project or offer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Project or offer state does not allow accepting the offer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{projectId}/offers/{offerId}/reject:
    post:
      tags:
        - Offers
      operationId: rejectOffer
      summary: Reject offer for project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/OfferId'
      responses:
        '200':
          description: Offer rejected successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferResponse'
        '403':
          description: User is not allowed to manage this project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project or offer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Offer state does not allow rejection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{projectId}/ratings:
    post:
      tags:
        - Ratings
      operationId: createRating
      summary: Rate freelancer for project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRatingRequest'
      responses:
        '201':
          description: Rating created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RatingResponse'
        '400':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project or user profile not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Project is not completed or target user is not a freelancer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    OfferId:
      name: offerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CreateProjectRequest:
      type: object
      required:
        - title
        - description
        - budget
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          minLength: 1
          maxLength: 5000
        budget:
          type: number
          format: double
          minimum: 0.01
    UpdateProjectRequest:
      type: object
      required:
        - title
        - description
        - budget
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          minLength: 1
          maxLength: 5000
        budget:
          type: number
          format: double
          minimum: 0.01
    ProjectResponse:
      type: object
      required:
        - id
        - clientId
        - title
        - description
        - budget
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        assignedFreelancerId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          minLength: 1
          maxLength: 5000
        budget:
          type: number
          format: double
        status:
          type: string
          enum:
            - OPEN
            - IN_PROGRESS
            - COMPLETED
            - CANCELLED
        createdAt:
          type: string
          format: date-time
    UpdateProfileRequest:
      type: object
      required:
        - displayName
        - bio
        - skills
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 255
        bio:
          type: string
          maxLength: 5000
        skills:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 100
          uniqueItems: true
    UserProfileResponse:
      type: object
      required:
        - id
        - email
        - displayName
        - bio
        - role
        - skills
        - averageRating
        - ratingCount
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        bio:
          type: string
        role:
          type: string
          enum:
            - CLIENT
            - FREELANCER
            - ADMIN
        skills:
          type: array
          items:
            type: string
        averageRating:
          type: number
          format: double
        ratingCount:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
    FreelancerResponse:
      allOf:
        - $ref: '#/components/schemas/UserProfileResponse'
    CreateOfferRequest:
      type: object
      required:
        - freelancerId
        - amount
        - message
      properties:
        freelancerId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
          minimum: 0.01
        message:
          type: string
          minLength: 1
          maxLength: 5000
    OfferResponse:
      type: object
      required:
        - id
        - projectId
        - freelancerId
        - amount
        - message
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        freelancerId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        message:
          type: string
        status:
          type: string
          enum:
            - PENDING
            - ACCEPTED
            - REJECTED
            - CANCELLED
        createdAt:
          type: string
          format: date-time
    CreateRatingRequest:
      type: object
      required:
        - clientId
        - freelancerId
        - score
        - comment
      properties:
        clientId:
          type: string
          format: uuid
        freelancerId:
          type: string
          format: uuid
        score:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
        comment:
          type: string
          minLength: 1
          maxLength: 5000
    RatingResponse:
      type: object
      required:
        - id
        - projectId
        - clientId
        - freelancerId
        - score
        - comment
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        freelancerId:
          type: string
          format: uuid
        score:
          type: integer
          format: int32
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time